pyproject.toml
[tool.poetry]
name = "agfirst"
version = "0.1.0"
description = "Django backend for the AgFirst App"
authors = ["Lorence Lim <jlorencelim@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.0.3"
djangorestframework = "^3.14.0"

[tool.poetry.dev-dependencies]
factory-boy = "^3.3.0"
pytest = "^7.4.3"
pytest-django = "^4.7.0"
pytest-factoryboy = "^2.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


# pytest
[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "codeleap_backend.settings"
